+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.1.7]
 
 - Fixed USD value formatting and rate display.
+- Added Evm and Sui support for Backpack wallet.
 
 ## [1.1.6]